Updated Dashabord w/ Features
The dashboard now actually serves a purpose.

* Admins can now approve/unapprove elections as
election commissioners create them. If an admin
(head of student orgs) unapproves a ballot,
it is only hidden from student/public view. For
a ballot to be removed completely, a SuperAdmin
(system admins) must purge it from the system with
their interface in the dashboard (TODO).

* Admins can now view a listing of all users and
the information on said users.

* Admins can now individually approve and remove
election commissioners.