Moved admin-specific JavaScript to seperate file.
Keep in mind, this is in no way for security.
Security by obscurity - especially in an open
source project like this - is a terrible tactic.
Also, as stated in the README, you should assume
this project/prototype contains no proper security
measures.

The real reason for moving the scripts to a
separate location is for organizational purposes
and to save resources so end users aren't loading
scripts they don't need to.